typescript conversions
